+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Chat Patches `8.0-alpha.3` for Minecraft 1.20.1-1.21.8 on Fabric, Quilt
+- Fixed the search bar not initially working with certain settings
+- Fixed the context menu using incorrect messages when searching
+
 ## Chat Patches `8.0-alpha.2` for Minecraft 1.20.1-1.21.8 on Fabric, Quilt
 ### Since alpha.1:
 - [#243](https://www.github.com/mrbuilder1961/ChatPatches/issues/243)